Core app removed ❌
