--- applications/inventory/choices.py
@@ -1,6 +1,0 @@
-INVENTORY_STATE = [
-    ('RFS', 'Ready for Sale'),
-    ('RSV', 'Reserved'),
-    ('SLD', 'Sold'),
-    ('NFS', 'Not for Sale'),
-]
