+++ applications/products/admin.py
