+++ applications/products/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class InventoryConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'applications.inventory'
+    name = 'applications.products'
