+++ applications/supply/choices.py
@@ -8,3 +8,11 @@ ORDER_STATES = [
     ('draft', 'Borrador'),
     ('cancelled', 'Cancelada'),
 ]
+
+# Inventory product states
+INVENTORY_STATE = [
+    ('RFS', 'Ready for Sale'),
+    ('RSV', 'Reserved'),
+    ('SLD', 'Sold'),
+    ('NFS', 'Not for Sale'),
+]
